refactor(spring-security) : 스프링 시큐리티 적용

+++ src/main/java/ict/board/controller/ReplyController.java
@@ -1,11 +1,10 @@
 package ict.board.controller;
 
-import ict.board.config.annotation.Login;
-import ict.board.config.annotation.OnlyIctAndAuthor;
-import ict.board.config.argumentresolver.LoginMemberArgumentResolver.LoginSessionInfo;
 import ict.board.dto.ReplyForm;
 import ict.board.service.ReplyService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -22,21 +21,31 @@ public class ReplyController {
 
     @PostMapping("/board/{boardId}/reply")
     public String addReply(@PathVariable Long boardId, @ModelAttribute ReplyForm replyForm,
-                           @Login LoginSessionInfo loginSessionInfo) {
-        replyService.saveByReplyForm(replyForm, boardId, loginSessionInfo.getEmail());
+                           @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return "redirect:/login";
+        }
+        String email = userDetails.getUsername();
+        replyService.saveByReplyForm(replyForm, boardId, email);
         return "redirect:/board/" + boardId;
     }
 
     @PostMapping("/board/{boardId}/reply/delete/{replyId}")
     public String deleteReply(@PathVariable Long replyId, @PathVariable Long boardId,
-                              @Login LoginSessionInfo loginSessionInfo, Model model) {
+                              @AuthenticationPrincipal UserDetails userDetails, Model model) {
+        if (userDetails == null) {
+            return "redirect:/login";
+        }
         replyService.deleteReply(replyId);
         return "redirect:/board/" + boardId;
     }
 
     @PutMapping("/board/{boardId}/reply/update/{replyId}")
     public String updateReply(@RequestParam String content, @PathVariable Long replyId, @PathVariable Long boardId,
-                              @Login LoginSessionInfo loginSessionInfo) {
+                              @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return "redirect:/login";
+        }
         replyService.updateReply(replyId, content);
         return "redirect:/board/" + boardId;
     }
